+++ src/node/desktop/src/main/gwt-callback.ts
@@ -385,6 +385,7 @@ export class GwtCallback extends EventEmitter {
       // discover available R installations
       const rInstalls = findRInstallationsWin32();
       if (rInstalls.length === 0) {
+        logger().logErrorMessage('No R installations found via registry or common R install locations.');
         return '';
       }